Set MAC on NICs
In order to run 2 kvms at the same time, and have them both connect to the
system vde (or any other shared-bus network), they need to have different
MAC addresses.  Add a script (kvm-macaddr) to generate a random MAC, and use it
to set the MAC on the NIC if it's not set in the conf file.